pyproject.toml
[project]
name = "sherman"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9",
    "aiosqlite>=0.19",
    "apluggy>=1.1.2",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[project.scripts]
sherman = "sherman.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
